Move StatusIndicator to frontend folder for ease of access and update references
